+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `Attention` base class, `MultiHeadAttention`, and `ScaledDotProductAttention` classes.
 - Add `branch_network` and `trunk_network` arguments to `DeepONet` to allow for custom network architectures.
 - Add `MaskedOperator` base class.
+- Add `DeepCatOperator`.
 
 ## 0.1.0